Use `fzf` to fuzzily select git branches
Fish has similar capability. You can type `git checkout <search>` and
then <TAB> to get a list of branches that match. I don't think it's quite
as easy to use as this is though.

This "implementation" uses `xargs` to get around a bug in Fish with
subcommands. I don't know why it works :)