Fix `fzf-cd-widget` ALT-C keybinding
This is a workaround because the offical fix is to change ALT-C to send
Meta, but we can't do that because we will lose the ability to type a #
symbol.

The other workaround is use use ESC-C, but we can't use that either
because we're using Vim keybindings as when ESC is pressed we go out of
insert mode.

https://github.com/junegunn/fzf/issues/164