Improve Gitsigns Git blame message display
We introduce a small delay so the log message takes longer to appear,
this makes navigating the file less fussy as log lines are not appearing
all the time.

We also move the display to the end of the line, instead of right
aligned I do prefer the look of right aligned but it was covering up the
code and getting in the way all the time.