Use `nvim-colorizer` to colorize "colour" strings
This works with Tailwind colours too, and is apparently faster...

For Elixir files with markup in them we change the display of colours to
colour the foreground (the actual text) instead of the background. This
is primarily for the benefit of Tailwind classes in `class` attributes
where the coloured backgrounds are very jarring.