Switch from Telescope to Snacks.picker
We have carried over the existing Telescope keybindings (for muscle
memory) and mapped them to what we think are the equivilant
Snacks.pickers.

We configure the pickers to show the "score" by default for all pickers
so we can better understand why a file comes where it comes in the list
of results -- might remove later.

`fzf` search syntax is supported as it was before with Telescope, so we
copy over the relevant documentation. (I forget this even exists, but it
is clearly useful to be able to search more accurately.)

We also calm-down our Lua diagnostics by including Snacks as one of the
globals to ignore.